git-svn-id: svn://localhost/public/MoSync/branches/release-2.4@2007 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ runtimes/java/platforms/androidJNI/buildJava.rb
@@ -16,20 +16,38 @@
 
 require 'fileutils'
 
+#
+# Preprocess the .jpp files and writes them to their output location were they can then be compiled
+#
+
 def preprocess_android_file(src_file, src_dir, shared_dir, output_dir)
 	jtmp_file = src_file.gsub(/.jpp$/, ".jtmp")		
 	java_file = src_file.gsub(/.jpp$/, ".java")
 	
 	puts "Processing " + java_file
 	
-	#-D_ANDROID_BLUETOOTH
+	buildFlags = "-D_android"
+	
+	# Add the _ANDROID_BLUETOOTH flag when building for version-5 or higher 
+	androidVersion = Integer(ENV['MOSYNC_ANDROID_BLUETOOTH'])
+	if androidVersion >= 7
+		buildFlags << " -D_ANDROID_BLUETOOTH"
+	end 
 	
 	# Preprocess the jpp file into a jtmp file, sed fixes the output if any
-	system("xgcc -x c -E -o #{output_dir}#{jtmp_file} -D_android -I#{shared_dir} -Isrc" +
+	success = system("xgcc -x c -E -o #{output_dir}#{jtmp_file} #{buildFlags} -I#{shared_dir} -Isrc" +
 		" #{src_dir}#{src_file} 2>&1 | sed \"s/\\([a-zA-Z/]\\+\\)\\(.[a-zA-Z]\\+\\):\\([0-9]\\+\\):/\\1\\2(\\3):/\"")
 	
+	if (!success)
+		exit 1
+	end
+	
 	# Use sed to comment the lines which the proprocessor added to the file and save it as a java file
-	system("sed \"s/^# /\\/\\//\" < #{output_dir}#{jtmp_file} > #{output_dir}#{java_file}");
+	success = system("sed \"s/^# /\\/\\//\" < #{output_dir}#{jtmp_file} > #{output_dir}#{java_file}");
+	
+	if (!success)
+		exit 1
+	end
 	
 	FileUtils.remove "#{output_dir}#{jtmp_file}"
 end
@@ -49,18 +67,22 @@ puts "shared lib: " + shared_java_source
 
 out_dir = "AndroidProject/src/com/mosync/java/android/"
 
-# Preprocess all the jpp files and create java files from them
+# Preprocess the com.mosync.java.android files, these will later be transformed to the applications real package
+#
 Dir.foreach(android_source) {|x| 
-	if (x == "MoSync.jpp" || x == "MoSyncPanicDialog.jpp" ) # || x == "MoSyncView.jpp" || x == "MoSyncThread.jpp" || x == "ThreadPool.jpp" || x == "BigPhatError.jpp" )
+	if (x == "MoSync.jpp" || x == "MoSyncPanicDialog.jpp" || x == "TextBox.jpp") # || x == "MoSyncView.jpp" || x == "MoSyncThread.jpp" || x == "ThreadPool.jpp" || x == "BigPhatError.jpp" )
 		preprocess_android_file(x, "#{android_source}/", shared_java_source, out_dir)
 	end
 }
 
 out_dir = "AndroidProject/src/com/mosync/internal/android/"
 
-# Preprocess all the jpp files and create java files from them
+# Preprocess the com.mosync.internal.android files
+#
 Dir.foreach(android_source) {|x| 
 	if (x == "MoSyncView.jpp" || x == "MoSyncThread.jpp" || x == "ThreadPool.jpp" || x == "BigPhatError.jpp" )
 		preprocess_android_file(x, "#{android_source}/", shared_java_source, out_dir)
 	end
 }
+
+exit 0
